A simple C++ project
This Bank Management System is based on a C++ project. It is used to Keep the records of
clients, employee  in Bank. The bank management system is an application for maintaining a
personal account in a bank. The system provides access to the customer to create an account,
deposit and withdraws the cash from his account, also to view reports of all accounts present.